lingamr: LiNGAM Algorithms for Causal Discovery

R implementation of LiNGAM (Linear Non-Gaussian Acyclic Model) algorithms for causal discovery. Based on the Python implementation by Ikeuchi et al. (2023) https://github.com/cdt15/lingam. The VAR-LiNGAM residual diagnostics are inspired by the VARLiNGAM R code of Moneta et al. https://sites.google.com/site/dorisentner/publications/VARLiNGAM.

Details

tutorial: vignette("lingamr", package = "lingamr")

References

Shimizu, S., et al. (2011). DirectLiNGAM: A direct method for learning a linear non-Gaussian structural equation model. Journal of Machine Learning Research, 12, 1225-1248.

Hyvärinen, A., Zhang, K., Shimizu, S., & Hoyer, P. O. (2010). Estimation of a structural vector autoregression model using non-Gaussianity. Journal of Machine Learning Research, 11, 1709-1731.

Moneta, A., Entner, D., Hoyer, P. O., & Coad, A. (2013). Causal inference by independent component analysis: Theory and applications. Oxford Bulletin of Economics and Statistics, 75(5), 705-730. VARLiNGAM R code: https://sites.google.com/site/dorisentner/publications/VARLiNGAM

Python implementation (DirectLiNGAM, VAR-LiNGAM): cdt15/lingam, https://github.com/cdt15/lingam

See also

Author

Maintainer: Osamu Morimoto galactic.supermarket@gmail.com

Authors:

Osamu Morimoto galactic.supermarket@gmail.com

Other contributors:

T. Ikeuchi [copyright holder]

G. Haraoka [copyright holder]

M. Ide [copyright holder]

W. Kurebayashi [copyright holder]

S. Shimizu [copyright holder]